requirements.txt
Django>=5.0.0,<6.0.0
psycopg2>=2.8.6,<2.10
djangorestframework>=3.12.4,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4.0
dj-rest-auth>=6.0.0,<7.0.0
django-allauth>=0.63.6,<0.64.0
django-cors-headers>=4.4.0,<4.5.0
channels>=4.1.0,<4.2.0
daphne>=4.1.2,<4.2.0
Pillow>=8.2.0,<10.3.0
requests>=2.26.0,<3.0.0
